Added changes for minimal version
Will slowly re-add content as required
