+++ src/App.css
@@ -1,58 +1,13 @@
-:root {
-  --background-color: rgba(8, 14, 18, 1);
-  --background-fade: rgba(8, 14, 18, 0.75);
-  --theme-gold: #f2ae0a;
-  --theme-dark-green: rgba(16, 28, 36, 1);
-  --theme-background-green: rgba(0, 0, 0, 0.75);
-}
-
 .App {
   overflow-x: hidden;
-  z-index: 1;
   min-height: 100vh;
-  max-width: 100vw;
-  background-image: url("./assets/wallpaper2.jpg");
-  background: linear-gradient(rgba(0, 0, 0, 0.75), rgba(0, 0, 0, 0.75)),
-    url("./assets/wallpaper2.jpg");
-  background-repeat: no-repeat;
-  background-attachment: fixed;
-  background-position: center;
-  background-size: cover;
-  background-color: var(--background-color);
-  font-family: "Roboto", sans-serif;
+  min-width: 100vw;
+  background-color: black;
+  font-family: 'Ubuntu', sans-serif;
 }
 
 .section {
-  padding-top: 3%;
   min-height: 100vh;
-  max-width: 100vw;
-  z-index: 3;
-}
-
-.sectionTitle {
-  color: var(--theme-gold);
-  font-size: 65px;
-  text-align: center;
-  display: block;
+  min-width: 100vw;
   padding-top: 5%;
-  padding-bottom: 5%;
-}
-
-.sectionTitleMobile {
-  color: var(--theme-gold);
-  font-size: 50px;
-  text-align: center;
-  display: block;
-  padding-bottom: 10%;
-}
-
-.sectionBody {
-  margin-right: 10%;
-  margin-left: 10%;
-}
-
-@media screen and (max-width: 500px) {
-  .sectionTitle {
-    font-size: 50px;
-  }
 }
