--- src/components/about/about.css
@@ -1,47 +1,0 @@
-.aboutSection {
-  padding-top: 3%;
-  padding-bottom: 3%;
-  max-width: 100vw;
-  min-height: 100vh;
-  z-index: 2;
-}
-
-.aboutSectionBody {
-  margin-right: 15%;
-  margin-left: 15%;
-}
-
-.aboutDescription {
-  color: white;
-  font-size: 26px;
-  text-align: center;
-  padding-top: 5%;
-  padding-bottom: 10%;
-}
-
-.aboutLinkLabelContainer {
-  display: flex;
-  justify-content: flex-end;
-}
-
-.aboutLinkLabel {
-  color: whitesmoke;
-  text-align: right;
-  font-size: 20px;
-}
-
-.avatarLinks {
-  width: 100vw;
-  text-align: center;
-}
-
-.aboutLink {
-  font-size: 20px;
-  text-align: left;
-}
-
-@media screen and (max-width: 500px) {
-  .aboutDescription {
-    font-size: 16px;
-  }
-}
