--- src/components/footer/footer.css
@@ -1,10 +1,0 @@
-.footer {
-  margin-right: 10%;
-  margin-left: 10%;
-}
-
-#footerSocialButton {
-  color: white;
-  font-size: 30px;
-  margin-bottom: 15%;
-}
