--- src/components/projects/projects.css
@@ -1,83 +1,0 @@
-.projectCardBody {
-  background-color: var(--theme-background-green);
-  height: 200px;
-}
-
-.projectCardRow {
-  padding-bottom: 3%;
-}
-
-.projectCardFooter {
-  padding-left: 5%;
-  padding-right: 5%;
-  padding-bottom: 3%;
-  background: var(--theme-background-green);
-}
-
-#cardTitle {
-  font-size: 20px;
-  color: var(--theme-gold);
-}
-
-.card-text {
-  color: whitesmoke;
-  font-size: 16px;
-}
-
-#githubIcon {
-  color: var(--theme-gold);
-  display: flex;
-  justify-content: flex-end;
-  font-size: 25px;
-}
-
-#githubIconNA {
-  color: var(--theme-dark-green);
-  display: flex;
-  justify-content: flex-end;
-  font-size: 25px;
-}
-
-.projectSkills {
-  font-family: "Cousine", monospace;
-  font-size: 12px;
-  color: var(--theme-gold);
-}
-
-@media screen and (max-width: 500px) {
-  .projectCardBody {
-    min-height: 10px;
-  }
-
-  #cardTitle {
-    font-size: 18px;
-  }
-
-  .card-text {
-    font-size: 12px;
-  }
-
-  .projectSkills {
-    font-size: 12px;
-  }
-
-  #githubIcon {
-    font-size: 20px;
-  }
-
-  #githubIconNA {
-    font-size: 20px;
-  }
-
-  .projectCardFooter {
-    padding-top: 5%;
-  }
-
-  .projectCard {
-    padding-bottom: 5%;
-  }
-
-  .projectCardRow {
-    padding-bottom: 0%;
-  }
-}
