--- src/components/skills/skills.css
@@ -1,70 +1,0 @@
-.skill {
-  justify-content: center;
-  color: white;
-}
-
-.skillRow {
-  padding-left: 5%;
-}
-
-.skillSection {
-  justify-content: center;
-  padding-top: 5%;
-}
-
-.skillListSection {
-  padding-bottom: 5%;
-  padding-left: 5%;
-  text-align: center;
-}
-
-.skillIcon {
-  color: white;
-  font-size: 25px;
-}
-
-.skillLabel {
-  font-size: 14px;
-  padding-top: 5%;
-  font-weight: 600;
-  color: var(--theme-gold);
-}
-
-.skillLabelMobile {
-  font-size: 14px;
-  padding-top: 5%;
-  color: whitesmoke;
-}
-
-.skillsListItem {
-  background-color: var(--background-color);
-  color: var(--theme-gold);
-  font-size: 25px;
-}
-
-.skillsListItemMobile {
-  background-color: var(--background-color);
-  color: var(--theme-gold);
-  font-size: 25px;
-  text-align: center;
-}
-
-.graphicsDemo {
-  justify-content: center;
-}
-
-.graphicsDemoDescription {
-  color: whitesmoke;
-  width: 100%;
-  text-align: center;
-}
-
-@media screen and (max-width: 500px) {
-  .skillRow {
-    padding-left: 0%;
-  }
-
-  .skillListSection {
-    padding-left: 0%;
-  }
-}
